feat: Allow for customization of the terminal presentation
This adds an additional layer to the terminal_provider config option
that takes a table instead of a string. If provided, the table will be
used as the configuration for the created terminal instance. The builtin
just supports minor changes such as altering the
width/height/border/style, but the Snacks terminal takes the full
configuration options for Snacks itself, providing for significantly
more control. We don't plan on supporting additional configuration for
the builtin, anyone wanting more intricate configuration should look at
using Snacks for the terminal provider.

Fixes #2