fix: Improve term::close() for Snacks terminal
Previously, it was possible for the closed state to get out of date,
resulting in the terminal opening instead of closing (due to the use of
term:toggle()). This updates the wrapper so that we only call toggle if
we're actually open, and to dynamically return the closed state instead
of using a static value.